tests added

+++ vllm_omni/diffusion/models/ovis_image/pipeline_ovis_image.py
@@ -366,8 +366,8 @@ class OvisImagePipeline(
 
         # VAE applies 8x compression on images but we must also account for packing which requires
         # latent height and width to be divisible by 2
-        height = 2 * (int(height) // (vae_scale_factor * 2))
-        width = 2 * (int(width) // (vae_scale_factor * 2))
+        height = int(2 * (int(height) // (vae_scale_factor * 2)))
+        width = int(2 * (int(width) // (vae_scale_factor * 2)))
 
         latents = latents.view(batch_size, height // 2, width // 2, channels // 4, 2, 2)
         latents = latents.permute(0, 3, 1, 4, 2, 5)
@@ -388,8 +388,8 @@ class OvisImagePipeline(
     ):
         # VAE applies 8x compression on images but we must also account for packing which requires
         # latent height and width to be divisible by 2.
-        height = 2 * (int(height) // (self.vae_scale_factor * 2))
-        width = 2 * (int(width) // (self.vae_scale_factor * 2))
+        height = int(2 * (int(height) // (self.vae_scale_factor * 2)))
+        width = int(2 * (int(width) // (self.vae_scale_factor * 2)))
 
         shape = (batch_size, num_channel_latents, height, width)
